+++ Release Notes.md
@@ -19,6 +19,7 @@ I've also adjusted the indentation of multiline function signature end parenthes
 * `Collection+EnumeratedArray` adds `enumeratedArray()` as a shorthand for `Array(self.enumerated())`. 
 * `DotAnimation` is a re-added, previously removed view, that animates a series of dots.
 * `DotAnimationText` is a re-added, previously removed view, that renders a text followed by a dot animation.
+* `ImagePicker` has new `allSourceTypes` and `availableSourceTypes` properties.
 * `View+PersistentSystemOverlays` adds a view extension for only hiding persistent system overlays if the feature is available.
 * `View+Searchable` adds a conditional searchable view extension.